Refactor corpus utilities into SRP modules
